feat: add user deletion functionality with confirmation modal
- Implemented a new method in AdminUserController to permanently delete users, including handling related data and session cleanup.
- Added a ConfirmDeleteUserModal component for user confirmation before deletion.
- Updated UsersIndex and UserShow pages to integrate the delete functionality, allowing admins to delete users with appropriate checks and feedback.
- Enhanced user experience by providing alerts for successful or error outcomes during deletion.

+++ app/Http/Controllers/Api/RemoteConfigController.php
@@ -21,7 +21,7 @@ class RemoteConfigController extends ApiController
             ?? rtrim((string) config('app.url'), '/') . '/api/v1';
 
         return $this->success([
-            'version'  => (int) config('remote_config.version', 1),
+            'version'  =>  config('remote_config.version', 1),
             'presensi' => [
                 'base_url'         => $presensiBaseUrl,
                 'timeout_seconds'  => (int) config('remote_config.presensi.timeout_seconds', 30),
